# apps/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelf API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "def-3": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "retryable", "requestId"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTHENTICATION_REQUIRED",
                  "AUTHORIZATION_DENIED",
                  "INVALID_REQUEST",
                  "IDEMPOTENCY_CONFLICT",
                  "ARTIFACT_NOT_FOUND",
                  "REVISION_NOT_FOUND",
                  "SHARE_NOT_FOUND",
                  "ACCESS_CREDENTIAL_NOT_FOUND",
                  "RANGE_NOT_SATISFIABLE",
                  "MULTI_RANGE_UNSUPPORTED",
                  "REQUEST_CANCELLED",
                  "CONTENT_UNAVAILABLE",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ],
                "title": "ErrorCode"
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "retryable": {
                "type": "boolean"
              },
              "requestId": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["reason"],
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "reason": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 256
                    }
                  },
                  "additionalProperties": false,
                  "title": "ErrorDetail"
                },
                "maxItems": 32
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false,
        "title": "ErrorEnvelope"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/artifacts/{artifactId}/shares": {
      "post": {
        "operationId": "createShareV1",
        "summary": "Create an unlisted latest or pinned share",
        "tags": ["shares"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["target"],
                "properties": {
                  "target": {
                    "type": "object",
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": ["latest", "pinned"]
                      },
                      "revisionId": {
                        "type": "string",
                        "pattern": "^rev_[A-Za-z0-9_-]{22}$"
                      }
                    },
                    "required": ["mode"],
                    "additionalProperties": false,
                    "allOf": [
                      {
                        "if": {
                          "properties": {
                            "mode": {
                              "enum": ["pinned"]
                            }
                          },
                          "required": ["mode"]
                        },
                        "then": {
                          "required": ["revisionId"]
                        },
                        "else": {
                          "not": {
                            "required": ["revisionId"]
                          }
                        }
                      }
                    ]
                  },
                  "expiresAt": {
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^art_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "artifactId",
            "required": true,
            "description": "Opaque artifact identifier containing 128 bits of entropy."
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-8"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/shares": {
      "get": {
        "operationId": "listSharesV1",
        "summary": "List share management records without capability material",
        "tags": ["shares"],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-9"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/shares/{shareId}": {
      "delete": {
        "operationId": "revokeShareV1",
        "summary": "Idempotently revoke one share",
        "tags": ["shares"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^shr_[A-Za-z0-9_-]{22}$"
            },
            "in": "path",
            "name": "shareId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/shares/{shareId}/resolve": {
      "post": {
        "operationId": "resolvePublicShareV1",
        "summary": "Resolve an unlisted share capability",
        "tags": ["public shares"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["secret"],
                "properties": {
                  "secret": {
                    "type": "string",
                    "maxLength": 512
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "shareId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-10"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/shares/{shareId}/content": {
      "post": {
        "operationId": "downloadPublicShareContentV1",
        "summary": "Download exact shared file bytes as an attachment",
        "tags": ["public shares"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["secret"],
                "properties": {
                  "secret": {
                    "type": "string",
                    "maxLength": 512
                  }
                },
                "additionalProperties": false
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["secret"],
                "properties": {
                  "secret": {
                    "type": "string",
                    "maxLength": 512
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "shareId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Exact immutable file bytes, always delivered as an attachment.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Always no-store."
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "Safe attachment file name."
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                },
                "description": "Complete immutable byte count."
              },
              "Referrer-Policy": {
                "schema": {
                  "type": "string"
                },
                "description": "Always no-referrer."
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string"
                },
                "description": "Always nosniff."
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Exact immutable file bytes, always delivered as an attachment."
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "499": {
            "description": "Default Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            }
          }
        }
      }
    },
...
  }
}